# dtlaunch: remember page between dtlaunch instances