--- a/apps/dtlaunch/settings-b2.js
+++ b/apps/dtlaunch/settings-b2.js
@@ -8,6 +8,7 @@
     swipeExit: false,
     timeOut: "Off",
     interactionBuzz: false,
+    rememberPage: false,
   }, require('Storage').readJSON(FILE, true) || {});
 
   function writeSettings() {
@@ -61,6 +62,13 @@
       value: settings.interactionBuzz,
       onchange: v => {
         settings.interactionBuzz = v;
+        writeSettings();
+      }
+    },
+    /*LANG*/'Remember Page': {
+      value: settings.rememberPage,
+      onchange: v => {
+        settings.rememberPage = v;
         writeSettings();
       }
     },