@bobrippling interested what you think about this!

Available on https://thyttan.github.io/BangleApps/?id=dtlaunch

- [x] test for ram leaks