@thyttan 2024-11-05
hm, the linter doesn't seem to handle `??` well.