@thyttan 2024-11-05
It feels like this adds just enough milliseconds to be noticable when switching back and forth to the clock. 

Might do some timings to compare.