@thyttan 2024-11-18
Measuring performance on non-pretokenized code:

Time to first draw: 438 ms -> 441 ms

Time to run whole app: 600 ms -> 615 ms